requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
PyJWT==2.8.0
bcrypt==4.1.2
azure-cosmos==4.5.1
azure-storage-blob==12.19.0
python-dotenv==1.0.0
werkzeug==3.0.1
email-validator==2.1.0
gunicorn==21.2.0
